[Build Speed] Remove NodeInlines.h from LiveNodeList.h rdar://159675881 https://bugs.webkit.org/show_bug.cgi?id=298241
Reviewed by Andy Estes.

* Source/WebCore/dom/ClassCollection.h:
* Source/WebCore/dom/LiveNodeList.h:
(WebCore::LiveNodeList::invalidateCache const): Deleted.
(WebCore::LiveNodeList::document const): Deleted.
(WebCore::LiveNodeList::protectedDocument const): Deleted.
* Source/WebCore/dom/LiveNodeListInlines.h:
(WebCore::LiveNodeList::invalidateCache const):
(WebCore::LiveNodeList::document const):
(WebCore::LiveNodeList::protectedDocument const):

Canonical link: https://commits.webkit.org/299469@main
